refactor(game-victory): change helper to service
